tests/wpt/web-platform-tests/performance-timeline/tentative/include-frames-one-remote-child.sub.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
</body>
<script>
promise_test(() => {
    return new Promise(resolve => {
        performance.clearResourceTimings()

        // Create Remote child iframe.
        const childFrame = document.createElement('iframe')
        childFrame.src = "http://{{hosts[][www]}}:{{ports[http][0]}}/resources/child-frame.html"
        document.body.appendChild(childFrame)

        childFrame.addEventListener("load", () => {
            const entries = performance.getEntries(true)
            const navigationEntries = performance.getEntriesByType("navigation", true)
            const markedEntries = performance.getEntriesByName("entry-name", undefined, true)

            // 3 entries for parent, 0 for child.
            assert_equals(entries.length, 3)

            // 1 entry for parent.
            assert_equals(navigationEntries.length, 1)

            // 0 entries.
            assert_equals(markedEntries.length, 0)

            resolve()
        })
    })
}, "GetEntries of a parent Frame with one Cross-Origin child")
</script>
